Remove filePrefix, bump version number.
It's probably better if we work based on what the first char of a string is.
This leads to less clutter and, while it might be a bit harder to extend
functionality, the author considers this trade-off worth it.

As for changes, see the new 'example.toml' file for an updated "syntax".